+++ src/StardewModdingAPI/IModHelper.cs
@@ -12,6 +12,9 @@
         /// <summary>Simplifies access to private game code.</summary>
         IReflectionHelper Reflection { get; }
 
+        /// <summary>Metadata about loaded mods.</summary>
+        IModRegistry ModRegistry { get; }
+
 
         /*********
         ** Public methods