+++ release-notes.md
@@ -7,6 +7,7 @@ For players:
 * Added mod folder path to the console output.
 
 For mod developers:
+* Added a mod registry which provides metadata about loaded mods (see `helper.ModRegistry`).
 * Fixed `SaveEvents.BeforeSave` and `SaveEvents.AfterSave` not triggering on days when the player shipped something.
 * Fixed `PlayerEvents.LoadedGame` and `SaveEvents.AfterLoad` being fired before the world finishes initialising.
 * Fixed some `LocationEvents`, `PlayerEvents`, and `TimeEvents` being fired during game startup.